Properly fix the environment check
For some reason the old environment variable would be unset by systemd
or something else, but not always, and not consistently. This would
result in a shell error as the variable would expand to nothing,
resulting in the expression `if test = ...`, which is invalid.

To work around this we simply check if the "gdm" process is running.
This should work regardless of what environment variables are present,
and doesn't rely on expanding any variables.